fix: problems with line spacing on comment lines
If you had this in an acrylic file:

    this is one line
    %% this is a comment

    this is another line

The final HTML render wouldn't have the space between the two lines that
aren't comments. This was happening because the instruction given to
comment-lines was to just not output anything, which in turn woudln't
even output the desired spacing.

I fixed this by reworking the spacing to be a separate `div`, instead of begin an attribute of each line (what was previously the HTML classes `line-normal` and `line-spaced`)